Refactor title, info, and attribute map out
