+++ website/src/lib/components/shared/index.ts
@@ -1,0 +1,2 @@
+export { default as Info } from "./info.svelte";
+export { default as Title } from "./title.svelte";
