+++ website/src/lib/components/verse1/word_chart.svelte
@@ -7,7 +7,7 @@ map component used in verse 3 for visualizing geographic data
     import { defaultWords } from "./keyframe.svelte";
     import cloud from "d3-cloud";
     import { fade } from "svelte/transition";
-    import Title from "../shared/title.svelte";
+    import { Title, Info } from "$lib/components/shared";
 
     // chart limits
     const similarityExtent = extent(homeSimilarityData.values()) as [number, number];
@@ -24,6 +24,13 @@ map component used in verse 3 for visualizing geographic data
     let hoveredWord = $state("");
 
     // data
+    type WordPlacement = {
+        size: number;
+        x: number;
+        y: number;
+        rotate: number;
+        text: string;
+    };
     const data = $derived(
         Array.from(
             homeSimilarityData.keys().map((i) => {
@@ -34,22 +41,8 @@ map component used in verse 3 for visualizing geographic data
             })
         ).sort((a, b) => b.size - a.size)
     );
-    let cloudWords: {
-        size: number;
-        x: number;
-        y: number;
-        rotate: number;
-        text: string;
-    }[] = [];
-    let words = $state(
-        [] as {
-            size: number;
-            x: number;
-            y: number;
-            rotate: number;
-            text: string;
-        }[]
-    );
+    let cloudWords: WordPlacement[] = [];
+    let words: WordPlacement[] = $state([]);
 
     // layout
     const layout = $derived.by(() => {
@@ -87,39 +80,9 @@ map component used in verse 3 for visualizing geographic data
     <Title title="What Does Home Mean to You?" />
 
     <!-- info tooltip -->
-    <div
-        class="tooltip tooltip-left absolute right-5 bottom-5 z-10 h-fit w-fit rounded-full p-0 hover:cursor-pointer"
-        data-tip={"Hover over a word to see cosine similarity with 'Home'." +
-            "\nWords collected from Habitat for Humanity's 'What does home mean to you' page, and cosine similarity calculated using API Ninjas' Text Similarity API."}
-    >
-        <svg
-            fill="#000000"
-            xmlns="http://www.w3.org/2000/svg"
-            width="25px"
-            height="25px"
-            viewBox="0 0 488.484 488.484"
-            xml:space="preserve"
-        >
-            <g>
-                <g>
-                    <path
-                        d="M244.236,0.002C109.562,0.002,0,109.565,0,244.238c0,134.679,109.563,244.244,244.236,244.244
-                    c134.684,0,244.249-109.564,244.249-244.244C488.484,109.566,378.92,0.002,244.236,0.002z M244.236,413.619
-                    c-93.4,0-169.38-75.979-169.38-169.379c0-93.396,75.979-169.375,169.38-169.375s169.391,75.979,169.391,169.375
-                    C413.627,337.641,337.637,413.619,244.236,413.619z"
-                    />
-                    <path
-                        d="M244.236,206.816c-14.757,0-26.619,11.962-26.619,26.73v118.709c0,14.769,11.862,26.735,26.619,26.735
-                    c14.769,0,26.62-11.967,26.62-26.735V233.546C270.855,218.778,259.005,206.816,244.236,206.816z"
-                    />
-                    <path
-                        d="M244.236,107.893c-19.949,0-36.102,16.158-36.102,36.091c0,19.934,16.152,36.092,36.102,36.092
-                    c19.929,0,36.081-16.158,36.081-36.092C280.316,124.051,264.165,107.893,244.236,107.893z"
-                    />
-                </g>
-            </g>
-        </svg>
-    </div>
+    <Info
+        tooltip="Hover over a word to see cosine similarity with 'Home'. Words collected from Habitat for Humanity's 'What does home mean to you' page, and cosine similarity calculated using API Ninjas' Text Similarity API."
+    />
 
     <!-- cloud -->
     <svg
