+++ website/src/lib/components/verse4/scatter.svelte
@@ -2,6 +2,7 @@
     import { data, type CountyRaces } from "$lib/data";
     import { extent, scaleSqrt, scaleOrdinal, scaleLinear } from "d3";
     import { fade } from "svelte/transition";
+    import { Info } from "$lib/components/shared";
 
     // Chart Data
     const width = 780;
@@ -326,38 +327,7 @@
     </div>
 
     <!-- Chart Tooltip -->
-    <div
-        class="tooltip tooltip-left absolute right-5 bottom-5 rounded-full hover:cursor-pointer"
-        data-tip={"Select Races from Legend to Filter the Graph." +
-            "\nSelect the Same Race Again to Return Graph to Original State (Showing All Races)." +
-            "\n Hover Over Data Points for More In-Depth Information"}
-    >
-        <svg
-            fill="#000000"
-            xmlns="http://www.w3.org/2000/svg"
-            width="25px"
-            height="25px"
-            viewBox="0 0 488.484 488.484"
-            xml:space="preserve"
-        >
-            <g>
-                <g>
-                    <path
-                        d="M244.236,0.002C109.562,0.002,0,109.565,0,244.238c0,134.679,109.563,244.244,244.236,244.244
-                c134.684,0,244.249-109.564,244.249-244.244C488.484,109.566,378.92,0.002,244.236,0.002z M244.236,413.619
-                c-93.4,0-169.38-75.979-169.38-169.379c0-93.396,75.979-169.375,169.38-169.375s169.391,75.979,169.391,169.375
-                C413.627,337.641,337.637,413.619,244.236,413.619z"
-                    />
-                    <path
-                        d="M244.236,206.816c-14.757,0-26.619,11.962-26.619,26.73v118.709c0,14.769,11.862,26.735,26.619,26.735
-                c14.769,0,26.62-11.967,26.62-26.735V233.546C270.855,218.778,259.005,206.816,244.236,206.816z"
-                    />
-                    <path
-                        d="M244.236,107.893c-19.949,0-36.102,16.158-36.102,36.091c0,19.934,16.152,36.092,36.102,36.092
-                c19.929,0,36.081-16.158,36.081-36.092C280.316,124.051,264.165,107.893,244.236,107.893z"
-                    />
-                </g>
-            </g>
-        </svg>
-    </div>
+    <Info
+        tooltip="Select races from legend to toggle filters, and hover over data points for more in-depth info. Data collected from US Census Bureau, censusreporter.org, and Logan et al.'s Longitudinal Tract Data Base (2000) and compiled on Kaggle."
+    />
 </div>
